tools/requirements.txt
# Requirements for HakPak Pro Stripe Webhook Handler
flask>=2.3.0
stripe>=5.4.0
gunicorn>=20.1.0
python-dotenv>=1.0.0
